Refactoring and add more packages. Integrate levels request in ServerController
