+++ src/main/java/com/example/server/controller/ServersController.java
@@ -1,23 +1,26 @@
 package com.example.server.controller;
 
-import com.example.server.entities.Admin;
-import com.example.server.entities.User;
-import com.example.server.entities.UserRegistrationForm;
+import com.example.server.entities.levelModels.Level;
+import com.example.server.entities.userModels.UserRegistrationForm;
 import com.example.server.services.LevelsService;
 import com.example.server.services.UsersService;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+/**
+ *
+ * localhost:8080 -> swagger url in der applicationproperties
+ *
+ * Post ->      Neue Dinge erzeugen
+ *
+ * Get ->       Bestehende Dinge abfragen
+ *
+ * Put ->       Bestehe Dinge manipulieren
+ *
+ * Delete ->    Dinge löschen (zb Sessions Cookies bei Logout)
+ *
+ */
 
 @RestController
 public class ServersController {
@@ -27,108 +30,37 @@ public class ServersController {
     @Autowired
     private UsersService usersService;
 
-    @GetMapping("/api/levels")
-    public @ResponseBody ResponseEntity getLevels() {
-        return ResponseEntity.status(200).body(this.levelsService.getAll());
-    }
-
     @GetMapping("/api/users")
-    public @ResponseBody ResponseEntity getUsers() {
-        return ResponseEntity.status(200).body(this.usersService.getAll());
-    }
+    public @ResponseBody ResponseEntity getUsers() { return ResponseEntity.status(200).body(this.usersService.getAll()); }
 
-    @GetMapping("/api/userByID")
-    public @ResponseBody ResponseEntity getUserByID(Integer id) {
-        User user = usersService.getUserById(id);
-        System.out.println("Benutzerkram name :" + user.getUserName() + "password lulu" + user.getPassword() );
-        return ResponseEntity.status(200).body(this.usersService.getUserById(id));
-    }
+    @GetMapping("/api/userById")
+    public @ResponseBody ResponseEntity getUserByID(Integer id) { return ResponseEntity.status(200).body(this.usersService.getUserById(id)); }
 
-    @PostMapping("/api/userCreation")
+    @PostMapping("/api/createUser")
     public @ResponseBody ResponseEntity createUser(@RequestBody UserRegistrationForm user) {
-        if (false) { //Datenbankabfrage ob es den User bereits gibt
-            // if (!usersService.checkIfUserExists(user)) { //Datenbankabfrage ob es den User bereits gibt
+        if (usersService.checkIfUserExists(user) == true) {
             return ResponseEntity.status(409).build();
         } else {
-
+            //FIXME user muss erst "befüllt" werden bevor er in die DB kommt
             usersService.saveUser(user);
             return ResponseEntity.status(200).build();
         }
+    }
 
+    @GetMapping("/api/levels")
+    public @ResponseBody ResponseEntity getLevels() { return ResponseEntity.status(200).body(this.levelsService.getAll()); }
+
+    @GetMapping("/api/level")
+    public @ResponseBody ResponseEntity getLevel(int id) { return ResponseEntity.status(200).body(this.levelsService.getLevelById(id)); }
+
+    @PostMapping("/api/createLevel")
+    public @ResponseBody ResponseEntity createLevel(@RequestBody Level level) {
+        //TODO Abfrage ob schon vorhanden
+        levelsService.save(level);
+        return ResponseEntity.status(200).build();
     }
 
 
 
-    /**
-     * GetMapping Erklärung
-     * OHHH BOI LEGO > Playmobil.
-     *
-     * v             v /
-     * http://localhost:8080/users
-     *
-     * localhost:8080 -> swagger url in der applicationproperties
-     *
-     *
-     * Post ->      Neue Dinge erzeugen
-     *
-     * Get ->       Bestehende Dinge abfragen
-     *
-     * Put ->       Bestehe Dinge manipulieren
-     *
-     * Delete ->    Dinge löschen (zb Sessions Cookies bei Logout)
-     */
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-    // clientseitig
-
-    public class HTTPController {
-
-
-        public List<User> getUsers() throws Exception{
-            ObjectMapper objectMapper = new ObjectMapper();
-            List<User> list = new ArrayList<>();
-            return Arrays.asList(objectMapper.readValue(getConnection("/users"), User.class));
-
-
-
-        }
-
-        public String getConnection(String urlToRead) throws Exception{
-
-            StringBuilder stringBuilder= new StringBuilder();
-            BufferedReader rd = null;
-
-                URL url = new URL("http://localhost:8080" + urlToRead);
-                HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-                connection.setRequestMethod("GET");
-                String line;
-                while ((line = rd.readLine()) != null) {
-                    stringBuilder.append(line);
-                }
-                rd.close();
-
-                rd = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-
-
-            return stringBuilder.toString();
-
-        }
-
-
-
-    }
 
 }
