+++ src/main/java/com/example/server/entities/userModels/Admin.java
@@ -1,6 +1,6 @@
-package com.example.server.entities;
+package com.example.server.entities.userModels;
 
-public class Admin extends User{
+public class Admin extends User {
 
     public Admin(){
         super();
