+++ src/main/java/com/example/server/entities/userModels/Role.java
@@ -1,4 +1,4 @@
-package com.example.server.entities;
+package com.example.server.entities.userModels;
 
 public enum Role {
     ADMIN,
