+++ src/main/java/com/example/server/entities/userModels/UserRegistrationForm.java
@@ -1,9 +1,9 @@
-package com.example.server.entities;
+package com.example.server.entities.userModels;
 
 import lombok.Data;
 
 @Data
-public class UserRegistrationForm  {
+public class UserRegistrationForm extends User {
 
     private String emailAdress;
     private String password;
