+++ src/main/java/com/example/server/repositorys/UsersRepository.java
@@ -1,6 +1,6 @@
 package com.example.server.repositorys;
 
-import com.example.server.entities.User;
+import com.example.server.entities.userModels.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
